Vector Unbox, masking support, and more for Vector API
- Adds support to generate vptest and a node that represents these semantics. This
is used for mask checking methods "anyTrue" and "allTrue".
- Added enum bounds for Vector API methods to improve robustness in checking whether
a method is a Vector API method.
- Fixed several bugs related to wrong class being set for VectorBox. Namely, the
receiver type was always set as the class, but in reality, the return type was correct.
- Added some more support around masks and some casting.
- Added support for Vector unboxing: generation of vector phis and also actually unbox
from Vector object.
- Added flag to turn off unstable_if speculation named "DisableUnstableIfOpt". This is
useful when testing because it reduces chances of getting to runtime from deopt.
- Added disabled support for allocations for VectorBox. This is not yet enabled because
VectorBox nodes are removed but leave behind the allocations which adds overhead.
- For mask support, it was ensured that VectorBox and VectorUnbox also work with it.
- Invariant nodes should be hoistable now.
- Now support int128, int256, int512, float128, float256, float512, and double128,
double256, and double512.