Add support for generation of VectorBox
VectorBox is unconditionally placed in graph along with allocations
during the initial intrinsifications. However, VectorBox is considered
for elimination when its output object is not being used.

In order to also deal with VectorMask boxing, additional nodes have
been added to "convert" between internal representation of masks
to the array representation used in GenericMask.