Add support for mask ops and mask making
Added support for code generation for or/and operation on masks.

Also, fixed the issue related to intrinsification of calls to trueMask
and falseMask. Unfortunately, due to the way intrinsics dispatch works,
I needed to add an intrinsic for every class currently supported - this
should be cleaned up in future.